requirements.txt
flask
flask-cors
PyMuPDF
pandas
scipy
google-generativeai
gTTS
pygame
deep-translator
pillow
